third_party/blink/web_tests/fast/html/imports/import-normalize-resolve-1-expected.txt

CONSOLE WARNING: line 9: document.registerElement is deprecated and will be removed in M80, around February 2020. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE MESSAGE: line 8: e01
CONSOLE MESSAGE: line 8: e02
CONSOLE MESSAGE: line 8: e03
CONSOLE MESSAGE: line 8: e04
CONSOLE MESSAGE: line 8: e05
CONSOLE MESSAGE: line 8: e06
CONSOLE MESSAGE: line 8: e07
CONSOLE MESSAGE: line 8: e08
CONSOLE MESSAGE: line 8: e09
CONSOLE MESSAGE: line 8: e10
CONSOLE MESSAGE: line 8: e11
